requirements.txt
flask
requests
playwright
pandas
PyMuPDF
matplotlib
python-dotenv